bulkio: Move target resolution code into backupbase package.
A code movement refactoring which moves methods accessible
from outside of backupccl (currently ResolveTargetsToDescriptors) into
a separate backupbase package. Because of various dependencies,
most supporting functions, including system schema backup/restore
information are also moved into backupbase.

This is a pure code movement to avoid dependency cycles between
backupccl and changefeedccl. Now, all cross package dependencies
live in backupbase and changefeedbase respectively.

Release Notes: None
