+++ pkg/ccl/backupccl/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
         "schedule_exec.go",
         "show.go",
         "split_and_scatter_processor.go",
-        "system_schema.go",
         "targets.go",
     ],
     embed = [":backupccl_go_proto"],
@@ -29,6 +28,7 @@ go_library(
     deps = [
         "//pkg/base",
         "//pkg/build",
+        "//pkg/ccl/backupccl/backupbase",
         "//pkg/ccl/storageccl",
         "//pkg/ccl/utilccl",
         "//pkg/featureflag",
@@ -117,14 +117,13 @@ go_test(
         "restore_old_versions_test.go",
         "show_test.go",
         "split_and_scatter_processor_test.go",
-        "system_schema_test.go",
-        "targets_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":backupccl"],
     deps = [
         "//pkg/base",
         "//pkg/blobs",
+        "//pkg/ccl/backupccl/backupbase",
         "//pkg/ccl/kvccl",
         "//pkg/ccl/partitionccl",
         "//pkg/ccl/storageccl",
@@ -148,21 +147,16 @@ go_test(
         "//pkg/sql",
         "//pkg/sql/catalog",
         "//pkg/sql/catalog/catalogkv",
-        "//pkg/sql/catalog/dbdesc",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/lease",
-        "//pkg/sql/catalog/schemadesc",
         "//pkg/sql/catalog/systemschema",
         "//pkg/sql/catalog/tabledesc",
-        "//pkg/sql/catalog/typedesc",
         "//pkg/sql/execinfra",
         "//pkg/sql/execinfrapb",
-        "//pkg/sql/parser",
         "//pkg/sql/rowenc",
         "//pkg/sql/rowflow",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
-        "//pkg/sql/types",
         "//pkg/storage/cloud",
         "//pkg/storage/cloudimpl",
         "//pkg/testutils",
