+++ pkg/ccl/backupccl/backupbase/system_schema_test.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package backupccl
+package backupbase
 
 import (
 	"context"
@@ -26,14 +26,14 @@ func TestAllSystemTablesHaveBackupConfig(t *testing.T) {
 	defer log.Scope(t).Close(t)
 
 	ctx := context.Background()
-	tc := testcluster.StartTestCluster(t, singleNode, base.TestClusterArgs{})
+	tc := testcluster.StartTestCluster(t, 1, base.TestClusterArgs{})
 	defer tc.Stopper().Stop(ctx)
 	sqlDB := sqlutils.MakeSQLRunner(tc.Conns[0])
 
 	systemTableNames := sqlDB.QueryStr(t, `USE system; SELECT table_name FROM [SHOW TABLES];`)
 	for _, systemTableNameRow := range systemTableNames {
 		systemTableName := systemTableNameRow[0]
-		if systemTableBackupConfiguration[systemTableName].includeInClusterBackup == invalid {
+		if SystemTableBackupConfiguration[systemTableName].IncludeInClusterBackup == InvalidBackupInclusion {
 			t.Fatalf("cluster backup inclusion not specified for system table %s", systemTableName)
 		}
 	}
@@ -43,12 +43,12 @@ func TestConfigurationDetailsOnlySetForIncludedTables(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 	defer log.Scope(t).Close(t)
 
-	for systemTable, configuration := range systemTableBackupConfiguration {
-		if configuration.customRestoreFunc != nil {
+	for systemTable, configuration := range SystemTableBackupConfiguration {
+		if configuration.CustomRestoreFunc != nil {
 			// If some restore options were specified, we probably want to also
 			// include in in the set of system tables that are looked at by cluster
 			// backup/restore.
-			if optInToClusterBackup != configuration.includeInClusterBackup {
+			if OptInToClusterBackup != configuration.IncludeInClusterBackup {
 				t.Fatalf("custom restore function specified for table %q, but it's not included in cluster backups",
 					systemTable)
 			}
