+++ pkg/ccl/backupccl/backupbase/targets_test.go
@@ -6,7 +6,7 @@
 //
 //     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
 
-package backupccl
+package backupbase
 
 import (
 	"context"
@@ -255,7 +255,7 @@ func TestDescriptorsMatchingTargets(t *testing.T) {
 			}
 			targets := stmt.AST.(*tree.Grant).Targets
 
-			matched, err := descriptorsMatchingTargets(context.Background(),
+			matched, err := DescriptorsMatchingTargets(context.Background(),
 				test.sessionDatabase, searchPath, descriptors, targets)
 			if test.err != "" {
 				if !testutils.IsError(err, test.err) {
@@ -265,11 +265,11 @@ func TestDescriptorsMatchingTargets(t *testing.T) {
 				t.Fatal(err)
 			} else {
 				var matchedNames []string
-				for _, m := range matched.descs {
+				for _, m := range matched.Descs {
 					matchedNames = append(matchedNames, m.GetName())
 				}
 				var matchedDBNames []string
-				for _, m := range matched.requestedDBs {
+				for _, m := range matched.RequestedDBs {
 					matchedDBNames = append(matchedDBNames, m.GetName())
 				}
 				sort.Strings(test.expected)
