+++ pkg/ccl/changefeedccl/changefeed_stmt.go
@@ -17,7 +17,7 @@ import (
 	"sort"
 	"time"
 
-	"github.com/cockroachdb/cockroach/pkg/ccl/backupccl"
+	"github.com/cockroachdb/cockroach/pkg/ccl/backupccl/backupbase"
 	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/changefeedbase"
 	"github.com/cockroachdb/cockroach/pkg/ccl/utilccl"
 	"github.com/cockroachdb/cockroach/pkg/docs"
@@ -186,7 +186,7 @@ func changefeedPlanHook(
 		}
 
 		// This grabs table descriptors once to get their ids.
-		targetDescs, _, err := backupccl.ResolveTargetsToDescriptors(
+		targetDescs, _, err := backupbase.ResolveTargetsToDescriptors(
 			ctx, p, statementTime, &changefeedStmt.Targets)
 		if err != nil {
 			return errors.Wrap(err, "failed to resolve targets in the CHANGEFEED stmt")
